LayoutTests/fast/canvas/check-stale-putImageData-expected.txt

CONSOLE MESSAGE: a == b? false
CONSOLE MESSAGE: a == c? false
1, 3, and 4 should all show the same red square. 2 should show nothing.
